Update g-counter to use tokio
Not really a performance gain since there is so much synchronous work
and there is very little asynchronous work that can be done. It can be
made more efficient by only passing what is necessary to potentially
allow simultaneous reads and writes to the same object, but that is more
complicated than this needs to be.